backend/requirements.txt
# FastAPI and web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
starlette>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.12.0
psycopg2-binary>=2.9.9

# Authentication & Security
python-jose[cryptography]>=3.3.0
argon2-cffi>=23.1.0
pyotp>=2.9.0

# Validation & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP client for external APIs
httpx>=0.25.0
aiohttp>=3.9.0

# Rate limiting
slowapi>=0.1.9

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Logging
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
